Added WorkChain implement for NEB calculation
The VaspNEBWorkChain follows the same style of input for VaspCalculation.
In addition, added the functionality for restarting the calculation.
Note that this classes inherent from the aiida-core's BaseRestartWorkChain
instead of the implemention in the plugin repository.

Revised input keywords for Incar class to disable parser-level input
checking. Otherwise it is not possible to supply non-standard keywords
that are needed for running VTST NEB calcualtions.

At the momment, only native VASP NEB calculations are not supported.