Added force parsing from outcar files
For NEB calculations we need the forces for further analysis, but
they are not in the vasprun.xml files (which is broken). They have
to be parsed from the OUTCAR file. The output node is a single
Array node contains arrays like `forces_01`, `forces_02` etc...

The OUTCAR parser code is also refectored to better integrate with
the original intended framework.